pyproject.toml
[tool.poetry]
name = "simsapa"
version = "0.4.0-alpha.1"
description = "Simsapa Dhamma Reader - Buddhist suttas and Pali dictionary"
# pypi identification
authors = ["Gambhiro <profound.labs@gmail.com>"]
license = "GPLv3"

# NOTE: Version updates:
# - tool.poetry.version (above)
# - simsapa.db_version (below)
# - tool.briefcase.version (below)
# - simsapa/__init__.py SIMSAPA_APP_VERSION
# - simsapa/scripts/bootstrap.sh: release_info TOML version_tag
# - simsapa/assets/releases_fallback.json

[simsapa]
db_version = "0.4.0-alpha.1"

[tool.poetry.dependencies]
python = "~3.11"
PyQt6 = "~6.5.2"
# Check poetry.lock Qt version matches with briefcase workflow yml
PyQt6-Qt6 = "~6.5.2"
PyQt6-WebEngine = "~6.5.0"
PyQt6-WebEngine-Qt6 = "~6.5.2"
PyYAML = "~6.0.0"
SQLAlchemy = "~2.0.7"
SQLAlchemy-Utils = "~0.40.0"
alembic = "~1.10.2"
requests = "~2"
platformdirs = "^3.10.0"
# PyPDF2 = "~1.26.0"
# PyMuPDF = "~1"
epub_meta = "~0.0.7"
Pillow = "~9.3.0"
# pandas = "~1"
# openpyxl = "~3.0.10"
semver = "~2.13.0"
pyglossary = "~4"
lxml = "~4.9.2"
python-dotenv = "~0.21.0"
pyperclip = "~1.8.2"
networkx = "~2"
bokeh = "~2"
scipy = "~1.10.0"
# For building on MacOS 11
#scipy = { file = "scipy-1.10.0-cp311-cp311-macosx_11_0_arm64.whl" }
python-idzip = "~0.3.7"
typer = "~0.9.0"
click = "~8.1.7"
tomlkit = "~0.11.6"
feedparser = "~6.0.10"
#pyqtkeybind = [{ version = "~0.0.9", platform = "linux" },]
beautifulsoup4 = "~4.11.1"
MarkupSafe = "~2.1.1"
# Mako = "~1.2.4"
Flask = "~2.2.2"
Flask-Cors = "~3.0.10"
aksharamukha = "~2.0.4"
markdown = "^3.4.1"
psutil = "^5.9.4"
roman = "^4.1"
chardet = "^5.2.0"
urllib3 = "2.0.2"
datetime = "5.1"
deepmerge = "^1.1.0"

# NOTE: charset-normalizer = "^3.0.1" errors with PyInstaller
# ModuleNotFoundError: No module named 'charset_normalizer.md__mypyc'
# https://stackoverflow.com/questions/74182807/no-module-named-charset-normalizer-md-mypyc
charset-normalizer = "==2.1.0"
ebooklib = "^0.18"
openai = "^0.27.4"
tiktoken = "^0.3.3"

# 67.4.0 failed to install
setuptools = "65.7.0"
# 2022.7.1 failed to install
pytz = "2022.7.0"
# 2023.0 failed to install
pyinstaller-hooks-contrib = "2022.14"
# 2.3.7 failed to install
werkzeug = "2.3.6"
# 6.3.3 failed to install
tornado = "6.3.2"

tantivy = { git = "https://github.com/simsapa/tantivy-py.git", branch = "simsapa" }
# tantivy = { git = "https://github.com/simsapa/tantivy-py.git", branch = "simsapa", rev = "229c10213b32de3ad4c905c2e968798f693afc28" }
# tantivy = { path = "../libs-src/tantivy-py-simsapa", develop = true }
# tantivy = { path = "../libs-src/tantivy-py-simsapa/target/wheels/tantivy-0.20.2rc4-cp311-cp311-manylinux_2_34_x86_64.whl" }

[tool.poetry.scripts]
simsapa = 'simsapa.runner:main'

[tool.poetry.group.dev.dependencies]
pyinstaller = "5.13"
tuna = { git = "https://github.com/nschloe/tuna.git" }

# scalene = "^1.5.14"

# 7.2.1 failed to install
pytest = "7.2.0"
pyArango = "~2.0.0"
html5lib = "^1.1"
selenium = "^4.7.2"
dmgbuild = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"

[tool.ruff]
ignore = ["E401", "E501", "E712"]
target-version = "py311"

[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}

[tool.briefcase]
project_name = "Simsapa Dhamma Reader"
bundle = "io.github"
# briefcase requires PEP440 version strings
version = "0.4.0a1"
url = "https://github.com/simsapa/simsapa"
license = "MIT license"
# Package will display this info
author = 'Profound Labs'
author_email = "profound.labs@gmail.com"

[tool.briefcase.app.simsapa]
formal_name = "Simsapa Dhamma Reader"
description = "Buddhist suttas and Pali dictionary"
icon = "simsapa/assets/icons/appicons/simsapa"
sources = ['simsapa']

requires = [
"PyQt6              >= 6.5.2,    < 6.6.0",
"PyQt6-Qt6           == 6.5.2",
"PyQt6-WebEngine    >= 6.5.0,    < 6.6.0",
"PyQt6-WebEngine-Qt6 == 6.5.2",
"PyYAML             >= 6.0.0,    < 6.1.0",
"SQLAlchemy         >= 2.0.7,   < 2.1.0",
"SQLAlchemy-Utils   >= 0.40.0,   < 0.41.0",
"alembic            >= 1.10.2,    < 1.11.0",
"requests           >= 2.25.0,   < 3.0.0",
"platformdirs       >= 3.10.0,   < 4.0.0",
# "PyPDF2             >= 1.26.0,   < 1.27.0",
# "PyMuPDF            >= 1.18.0,   < 2.0.0",
"epub_meta          >= 0.0.7,    < 0.1.0",
"Pillow             >= 9.3.0,    < 9.4.0",
# "pandas             >= 1.2.0,    < 2.0.0",
# "openpyxl           >= 3.0.10,   < 3.1.0",
"semver             >= 2.13.0,   < 2.14.0",
"pyglossary         >= 4.0.0,    < 5.0.0",
"lxml               >= 4.9.2,    < 5.0.0",
"python-dotenv      >= 0.21.0,   < 0.22.0",
"pyperclip          >= 1.8.2,    < 1.9.0",
"networkx           >= 2.5.0,    < 3.0.0",
"bokeh              >= 2.4.0,    < 3.0.0",
"scipy              >= 1.10.0,   < 2.0.0",
"python-idzip       >= 0.3.7,    < 0.4.0",
"typer              >= 0.9.0,    < 0.10.0",
"click              >= 8.1.7,    < 9.0.0",
"tomlkit            >= 0.11.0,   < 0.12.0",
"feedparser         >= 6.0.10,   < 6.1.0",
"beautifulsoup4     >= 4.11.1,   < 4.12.0",
"MarkupSafe         == 2.1.1",
# "Mako               >= 1.2.4,    < 1.3.0",
"Flask              >= 2.2.2,    < 2.3.0",
"Flask-Cors         >= 3.0.10,   < 3.1.0",
"aksharamukha       >= 2.0.4,    < 2.1.0",
"markdown           >= 3.4.1,    < 3.5.0",
"psutil             >= 5.9.4,    < 5.10.0",
"roman              >= 4.1,      < 5.0",
"chardet            >= 5.2.0,    < 5.3.0",
"urllib3            == 2.0.2",
"datetime           == 5.1",
"deepmerge          == 1.1.0",
"charset-normalizer == 2.1.0",
"ebooklib           >= 0.18.0,   < 0.19.0",
"openai             >= 0.27.0,   < 0.28.0",
"tiktoken           >= 0.3.3,    < 0.4.0",
"setuptools         == 65.7.0",
"pytz               == 2022.7.0",
"pyinstaller-hooks-contrib == 2022.14.0",
"werkzeug           == 2.3.6",
"tornado            == 6.3.2",
"./tantivy-py/target/wheels/tantivy-0.20.2rc4-cp311-cp311-manylinux_2_34_x86_64.whl",
]

[tool.briefcase.app.simsapa.macOS]
requires = []

[tool.briefcase.app.simsapa.linux]
requires = [
#"pyqtkeybind >= 0.0.9, < 0.1.0",
]

system_requires = ["libtiff-dev"]

[tool.briefcase.app.simsapa.windows]
requires = []

# Mobile deployments
[tool.briefcase.app.simsapa.iOS]
requires = []

[tool.briefcase.app.simsapa.android]
requires = []
